Addressing event semantic changes that occured due to upgrading to d3 7.9.0.
When we migrated from ancient d3 to 7.9.0, we apparently did not do it
correctly. The new events handler now need to take an event and a
datum, in that order. Whereas in the previous versions, the order was
reversed.

This change addresses all mouseevent I can find. But I don't know
which ones have changed. So hopefully I did not miss any.

I was alerted to that because some of Saloni's changes fixed one of
those. And then realized "if it was wrong there, it could be wrong in
a lot of places".